+++ contracts/interfaces/IvSwapPoolDeployer.sol
@@ -21,8 +21,8 @@ interface IvSwapPoolDeployer {
             address factory,
             address token0,
             address token1,
-            uint24 fee,
-            uint24 vFee,
+            uint16 fee,
+            uint16 vFee,
             uint24 maxAllowListCount,
             uint256 maxReserveRatio
         );
